Split files. Sketch package. Work in progress.
