+++ Dockerfile
@@ -20,7 +20,7 @@ RUN wget -O rustup.sh https://sh.rustup.rs && \
     rm rustup.sh && \
     chmod -R 777 /rust
 
-COPY "./build_within_docker.py" "/build.py"
+COPY "build_contract_rust" "/build_contract_rust"
 
 ENV PATH="/rust/bin:${PATH}"
 ENV CARGO_HOME="/rust"
@@ -28,7 +28,7 @@ ENV RUSTUP_HOME="/rust"
 
 # Additional arguments (must be provided at "docker run"):
 # --no-wasm-opt (optional)
-ENTRYPOINT ["python", "./build.py", \
+ENTRYPOINT ["python", "/build_contract_rust/main.py", \
     "--project", "/project", \
     "--output", "/output", \
     "--cargo-target-dir", "/rust/cargo-target-dir"]
