+++ build_contract_rust/__init__.py
@@ -1,0 +1,3 @@
+from build_contract_rust import builder
+
+__all__ = ["builder"]
