+++ build_contract_rust/errors.py
@@ -1,0 +1,4 @@
+
+class ErrKnown(Exception):
+    def __init__(self, *args: object) -> None:
+        super().__init__(*args)
